--- src/app/articles/categories/[slug]/slug.server.test.ts
@@ -1,25 +1,0 @@
-// @jest-environment node
-
-import { mockCategories } from "mocks/data";
-import { server } from "mocks/msw/server";
-
-import { getStaticPaths } from "./page";
-
-beforeAll(() => server.listen());
-afterAll(() => server.close());
-
-describe("pages/articles/categories/[slug]/server", () => {
-  const mockCategoryList = Object.values(mockCategories);
-
-  it("getStaticPaths", async () => {
-    const { fallback, paths } = await getStaticPaths({});
-    expect(fallback).toBe("blocking");
-
-    const expectedPaths = mockCategoryList.map((category) => category.slug);
-
-    for (const path of paths) {
-      if (typeof path === "string") return;
-      expect(expectedPaths).toContain(path.params.slug);
-    }
-  });
-});
