+++ src/components/pages/categories/Category/index.tsx
